Import of v0.5.6

+++ ac3.h
@@ -26,11 +26,11 @@
 #endif
 
 typedef unsigned long long uint_64;
-typedef unsigned long  uint_32;
+typedef unsigned int uint_32;
 typedef unsigned short uint_16;
 typedef unsigned char  uint_8;
 
-typedef signed long    sint_32;
+typedef signed int sint_32;
 typedef signed short   sint_16;
 typedef signed char    sint_8;
